pyproject.toml
[project]
name = "commitgen"
version = "0.1.0"
description = "AI-powered Git commit message generator."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
commitgen = "commitgen.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
